+++ const_desctr_gc/object_control.c
@@ -1,28 +1,32 @@
 #include "object_control.h"
 
-void *default_construct(pool_allocator *allocator) {
-  return pool_alloc(allocator);
+void *default_construct(memory_context *mem_ctx, size_t size) {
+  return mem_ctx->allocate(mem_ctx->context, size);
 }
 
-void default_destruct(pool_allocator *allocator, void *object) {
-  pool_free(allocator, object);
+void default_destruct(memory_context *mem_ctx, void *object) {
+  mem_ctx->deallocate(object, mem_ctx->context);
 }
 
 int object_create(object_control *obj_control, size_t obj_size,
-                  void *(*construct)(pool_allocator *allocator),
-                  void (*destruct)(pool_allocator *allocator, void *object),
-                  pool_allocator *allocator) {
+                  void *(*construct)(memory_context *mem_ctx, size_t size),
+                  void (*destruct)(memory_context *mem_ctx, void *object),
+                  memory_context *mem_ctx) {
 
-  if (!obj_control || !allocator)
+  if (!obj_control || !mem_ctx)
     return MEMORY_ALLOC_ERROR;
 
-  if (obj_size == 0 || allocator->block_size < obj_size)
+  if (obj_size == 0)
     return MEMORY_SIZE_ERROR;
 
-  obj_control->allocator = allocator;
+  if (!mem_ctx->allocate || !mem_ctx->deallocate)
+    return MEMORY_ALLOC_ERROR;
+
+  obj_control->mem_ctx = mem_ctx;
   obj_control->construct = construct ? construct : default_construct;
   obj_control->destruct = destruct ? destruct : default_destruct;
-  obj_control->object = obj_control->construct(allocator);
+  mem_ctx->deallocate;
+  obj_control->object = obj_control->construct(obj_control->mem_ctx, obj_size);
 
   if (!obj_control->object)
     return MEMORY_ALLOC_ERROR;
@@ -34,9 +38,9 @@ int object_destroy(object_control *obj_control) {
   if (!obj_control)
     return MEMORY_ALLOC_ERROR;
 
-  obj_control->destruct(obj_control->allocator, obj_control->object);
+  obj_control->destruct(obj_control->mem_ctx, obj_control->object);
 
-  obj_control->allocator = NULL;
+  obj_control->mem_ctx = NULL;
   obj_control->construct = NULL;
   obj_control->destruct = NULL;
   obj_control->object = NULL;
