+++ const_desctr_gc/object_control.h
@@ -5,14 +5,20 @@
 #define MEMORY_SIZE_ERROR -2
 
 typedef struct {
+  void *(*allocate)(void *context, size_t size);
+  void (*deallocate)(void *ptr, void *context);
+  void *context;
+} memory_context;
+
+typedef struct {
   void *object;
-  void *(*construct)(pool_allocator *allocator);
-  void (*destruct)(pool_allocator *allocator, void *object);
-  pool_allocator *allocator;
+  void *(*construct)(memory_context *mem_ctx, size_t size);
+  void (*destruct)(memory_context *mem_ctx, void *object);
+  memory_context *mem_ctx;
 } object_control;
 
 int object_create(object_control *obj_control, size_t obj_size,
-                  void *(*construct)(pool_allocator *allocator),
-                  void (*destruct)(pool_allocator *allocator, void *object),
-                  pool_allocator *allocator);
+                  void *(*construct)(memory_context *mem_ctx, size_t size),
+                  void (*destruct)(memory_context *mem_ctx, void *object),
+                  memory_context *mem_ctx);
 int object_destroy(object_control *obj_control);
